Make separate folders for mkldnn codes test=develop
