+++ paddle/fluid/operators/mkldnn/CMakeLists.txt
@@ -1,0 +1,2 @@
+include(operators)
+register_operators()
