+++ python/paddle/fluid/tests/unittests/mkldnn/test_dequantize_mkldnn_op.py
@@ -16,35 +16,32 @@ from __future__ import print_function
 
 import unittest
 import numpy as np
-from op_test import OpTest
+from paddle.fluid.tests.unittests.op_test import OpTest
 
 
-class TestQuantizeOp(OpTest):
+class TestDeQuantizeOp(OpTest):
     def setUp(self):
-        self.op_type = 'quantize'
+        self.op_type = 'dequantize'
         self.scale = 2.0
         self.input_size = [1, 1, 5, 5]  #Naive nChw16c
-        self.is_negative = False
+        self.data_type = 'int8'
         self.set_scale()
-        self.set_is_negative()
+        self.set_data_type()
 
-        if self.is_negative:
-            input = (100 * np.random.random_sample(self.input_size) - 50
-                     ).astype('float32')
-            output = np.round(input * self.scale).astype('int8')
+        if self.data_type == 'int8':
+            input = (np.random.randint(0, 100, self.input_size) - 50
+                     ).astype(self.data_type)
+            output = (input * (1 / self.scale)).astype('float')
         else:
-            input = (100 *
-                     np.random.random_sample(self.input_size)).astype('float32')
-            output = np.round(input * self.scale).astype('uint8')
+            input = (np.random.randint(0, 100,
+                                       self.input_size)).astype(self.data_type)
+            output = (input * (1 / self.scale)).astype('float')
 
         self.inputs = {'Input': OpTest.np_dtype_to_fluid_dtype(input)}
 
         self.outputs = {'Output': output}
 
-        self.attrs = {
-            'Scale': self.scale,
-            'is_negative_input': self.is_negative
-        }
+        self.attrs = {'Scale': self.scale, }
 
     def test_check_output(self):
         self.check_output()
@@ -52,24 +49,24 @@ class TestQuantizeOp(OpTest):
     def set_scale(self):
         pass
 
-    def set_is_negative(self):
+    def set_data_type(OpTest):
         pass
 
 
-class TestQuantizeOp1(TestQuantizeOp):
+class TestDeQuantizeOp1(TestDeQuantizeOp):
     def set_scale(self):
         self.scale = 1.5
 
-    def set_is_negative(self):
-        self.is_nagative = True
+    def set_data_type(self):
+        self.data_type = 'int8'
 
 
-class TestQuantizeOp2(TestQuantizeOp):
+class TestDeQuantizeOp2(TestDeQuantizeOp):
     def set_scale(self):
-        self.scale = 0.1
+        self.scale = 0.8
 
-    def set_is_negative(self):
-        self.is_nagative = False
+    def set_data_type(self):
+        self.data_type = 'uint8'
 
 
 if __name__ == '__main__':
